requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.7.0
httpx>=0.26.0
openai>=1.12.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
click>=8.1.7
requests>=2.31.0
tabulate>=0.9.0
